+++ apps/pioneer/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @internal/playground
 
+## 0.2.7
+
+### Patch Changes
+
+- bump kk
+
+## 0.2.6
+
+### Patch Changes
+
+- bump kk
+
+## 0.2.5
+
+### Patch Changes
+
+- bump kk
+
+## 0.2.4
+
+### Patch Changes
+
+- bump hdwallet
+
+## 0.2.3
+
+### Patch Changes
+
+- bump kk wallet
+
+## 0.2.2
+
+### Patch Changes
+
+- deps keepkey wallet
+
+## 0.2.1
+
+### Patch Changes
+
+- y
+
 ## 0.2.0
 
 ### Minor Changes