+++ e2e/wallets/integration-wallet-package-keepkey/CHANGELOG.md
@@ -1,5 +1,82 @@
 # @coinmasters/integration-wallet-keepkey
 
+## 0.0.8
+
+### Patch Changes
+
+- bump kk
+
+- Updated dependencies []:
+  - @coinmasters/wallet-keepkey@8.0.7
+  - @coinmasters/types@4.2.7
+  - @coinmasters/core@6.0.7
+
+## 0.0.7
+
+### Patch Changes
+
+- bump kk
+
+- Updated dependencies []:
+  - @coinmasters/wallet-keepkey@8.0.6
+  - @coinmasters/types@4.2.6
+  - @coinmasters/core@6.0.6
+
+## 0.0.6
+
+### Patch Changes
+
+- bump kk
+
+- Updated dependencies []:
+  - @coinmasters/wallet-keepkey@8.0.5
+  - @coinmasters/types@4.2.5
+  - @coinmasters/core@6.0.5
+
+## 0.0.5
+
+### Patch Changes
+
+- bump hdwallet
+
+- Updated dependencies []:
+  - @coinmasters/wallet-keepkey@8.0.4
+  - @coinmasters/types@4.2.4
+  - @coinmasters/core@6.0.4
+
+## 0.0.4
+
+### Patch Changes
+
+- bump kk wallet
+
+- Updated dependencies []:
+  - @coinmasters/wallet-keepkey@8.0.3
+  - @coinmasters/types@4.2.3
+  - @coinmasters/core@6.0.3
+
+## 0.0.3
+
+### Patch Changes
+
+- deps keepkey wallet
+
+- Updated dependencies []:
+  - @coinmasters/wallet-keepkey@8.0.2
+  - @coinmasters/types@4.2.2
+  - @coinmasters/core@6.0.2
+
+## 0.0.2
+
+### Patch Changes
+
+- y
+
+- Updated dependencies []:
+  - @coinmasters/wallet-keepkey@8.0.1
+  - @coinmasters/types@4.2.1
+  - @coinmasters/core@6.0.1
+
 ## 1.2.0
 
 ### Minor Changes