+++ packages/swapkit/tokens/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @swapkit/tokens
 
+## 3.2.7
+
+### Patch Changes
+
+- bump kk
+
+## 3.2.6
+
+### Patch Changes
+
+- bump kk
+
+## 3.2.5
+
+### Patch Changes
+
+- bump kk
+
+## 3.2.4
+
+### Patch Changes
+
+- bump hdwallet
+
+## 3.2.3
+
+### Patch Changes
+
+- bump kk wallet
+
+## 3.2.2
+
+### Patch Changes
+
+- deps keepkey wallet
+
+## 3.2.1
+
+### Patch Changes
+
+- y
+
 ## 3.2.0
 
 ### Minor Changes